+++ src/components/Instances.hpp
@@ -7,6 +7,9 @@ static constexpr int32_t INSTANCES_INTERATE_OFFSET = 10;
 template <typename T>
 concept UObjectOrDerived = std::is_base_of_v<UObject, T>;
 
+using GNames_t   = TArray<FNameEntry*>*;
+using GObjects_t = TArray<UObject*>*;
+
 class InstancesComponent
 {
 public:
@@ -17,15 +20,17 @@ public:
 	void OnCreate();
 	void OnDestroy();
 
-	// initialize globals for RLSDK
+	bool initGlobals(); // initialize globals for RLSDK
+
+private:
 	uintptr_t FindPattern(HMODULE module, const unsigned char* pattern, const char* mask);
-	uintptr_t GetGNamesAddress();
-	uintptr_t GetGObjectsAddress();
-	uintptr_t getGMallocAddr();
-	bool      InitGlobals();
 	bool      AreGObjectsValid();
 	bool      AreGNamesValid();
 	bool      CheckGlobals();
+
+	uintptr_t findGNamesAddress();
+	uintptr_t findGMallocAddress();
+	uintptr_t findGPsyonixBuildIDAddress();
 
 private:
 	std::map<std::string, class UClass*>    m_staticClasses;